consolidated the _report_functions.py functions into _utils.py
